requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
joblib==1.4.2
requests==2.32.3
beautifulsoup4==4.12.3
sentence-transformers==3.0.1
plotly==5.24.1
streamlit==1.38.0
SQLAlchemy==2.0.35
psycopg[binary,pool]==3.2.1
redis==5.0.8
pinecone-client==5.0.1
google-generativeai==0.8.3
tavily-python
lightgbm
xgboost